feat(cli): add code-mod
