--- packages/cli/src/code-mod/mods/v0.0.142.d.ts
@@ -1,13 +1,0 @@
-/**
- * Convert useQueryFoo({}, {}, { staleTime: 5 }) to useQueryFoo({ params: {}, body: {},
- * options: { staleTime: 5 } })
- *
- * @param {InsightEvent} event
- * @param {boolean} verbose
- * @returns {Promise<void>}
- */
-export function executeCodeModVZeroDotZeroDotHundredFortyTwo(
-  event: InsightEvent,
-  verbose: boolean,
-): Promise<void>;
-//# sourceMappingURL=v0.0.142.d.ts.map
