--- packages/cli/src/code-mod/mods/v0.0.146.d.ts
@@ -1,13 +1,0 @@
-/**
- * Convert `queries.entitySelect(sql, { ...where })` to `queryEntity({ where: { ...where
- * } }).exec(sql)`
- *
- * @param {InsightEvent} event
- * @param {boolean} verbose
- * @returns {Promise<void>}
- */
-export function executeCodeModVZeroDotZeroDotHundredFortySix(
-  event: InsightEvent,
-  verbose: boolean,
-): Promise<void>;
-//# sourceMappingURL=v0.0.146.d.ts.map
